introduce wallet package
