+++ apps/extension/package.json
@@ -37,6 +37,7 @@
     "@repo/context": "workspace:*",
     "@repo/storage-chrome": "workspace:*",
     "@repo/ui": "workspace:*",
+    "@repo/wallet": "workspace:*",
     "@tanstack/react-query": "5.55.0",
     "exponential-backoff": "^3.1.1",
     "framer-motion": "^11.15.0",
