+++ apps/extension/src/state/persist.ts
@@ -4,7 +4,7 @@ import { produce } from 'immer';
 
 import { localExtStorage } from '@repo/storage-chrome/local';
 import { OriginRecord } from '@repo/storage-chrome/records';
-import { walletsFromJson } from '@repo/storage-chrome/wallet';
+import { Wallet } from '@repo/wallet';
 import { AppParameters } from '@penumbra-zone/protobuf/penumbra/core/app/v1/app_pb';
 
 export type Middleware = <
@@ -28,7 +28,7 @@ export const customPersistImpl: Persist = f => (set, get, store) => {
 
     set(
       produce((state: AllSlices) => {
-        state.wallets.all = walletsFromJson(wallets);
+        state.wallets.all = wallets.map(w => Wallet.fromJson(w));
         state.network.grpcEndpoint = grpcEndpoint;
         state.connectedSites.knownSites = knownSites as OriginRecord[];
         state.defaultFrontend.url = frontendUrl;
@@ -42,7 +42,7 @@ export const customPersistImpl: Persist = f => (set, get, store) => {
         const wallets = changes.wallets.newValue;
         set(
           produce((state: AllSlices) => {
-            state.wallets.all = walletsFromJson(wallets ?? []);
+            state.wallets.all = (wallets ?? []).map(w => Wallet.fromJson(w));
           }),
         );
       }
