--- packages/storage-chrome/src/encryption/index.ts
@@ -1,2 +1,0 @@
-export * from './box';
-export * from './key';
