+++ packages/storage-chrome/src/records/index.ts
@@ -1,3 +1,2 @@
 export type { OriginRecord } from './known-site';
 export { UserChoice } from './user-choice';
-export { Wallet, type WalletJson } from './wallet';
