+++ packages/wallet/package.json
@@ -1,0 +1,25 @@
+{
+  "name": "@repo/wallet",
+  "version": "0.0.1",
+  "private": true,
+  "license": "(MIT OR Apache-2.0)",
+  "type": "module",
+  "scripts": {
+    "lint": "eslint \"**/*.ts*\"",
+    "test": "vitest run"
+  },
+  "exports": {
+    ".": "./src/wallet.ts",
+    "./encryption": "./src/encryption/index.ts",
+    "./record": "./src/record.ts"
+  },
+  "dependencies": {
+    "@penumbra-zone/bech32m": "18.0.0",
+    "@penumbra-zone/protobuf": "11.0.0"
+  },
+  "devDependencies": {
+    "@penumbra-zone/types": "35.0.0",
+    "@penumbra-zone/wasm": "52.0.0",
+    "@types/chrome": "0.0.270"
+  }
+}
