+++ packages/wallet/src/encryption/index.ts
@@ -1,0 +1,2 @@
+export { Key, KeyPrint } from './key';
+export { Box } from './box';
