+++ packages/wallet/tsconfig.json
@@ -1,0 +1,7 @@
+{
+  "extends": "@repo/tsconfig/base.json",
+  "include": ["src", "vitest.config.ts", "tests-setup.ts"],
+  "compilerOptions": {
+    "lib": ["WebWorker", "DOM", "ESNext"]
+  }
+}
