+++ packages/wallet/vitest.config.ts
@@ -1,0 +1,3 @@
+import { defineConfig } from 'vitest/config';
+
+export default defineConfig({ test: { environment: 'jsdom', setupFiles: ['./tests-setup.ts'] } });
